refactor(server): move db create scripts to /bin; rename /db to /models
