+++ .travis.yml
@@ -31,8 +31,8 @@ before_script:
   # Copy configuration
   - cp config/local.json.travis config/local.json
   # Setup database
-  - server/db/create_db.sh postgres
-  - node server/db/create_test_data.js
+  - bin/create_db.sh postgres
+  - node bin/create_test_data.js
   # Configure and start xvfb
   - export DISPLAY=:99.0
   - sh -e /etc/init.d/xvfb start
