+++ README.md
@@ -25,9 +25,10 @@ To fetch dependencies and get cooking:
 1. `npm install`
 2. As part of the npm install process, the `postinstall` script will install the Bower dependencies for you.
 3. Copy `config/local.json.example` to `config/local.json`, and put your local info in there.
-4. Run `./server/db/create_db.sh` to create the database, tables, and indexes
+4. Run `./bin/create_db.sh` to create the database, tables, and indexes
   - this script currently hard-codes the db user, password, and dbname to 'chronicle' (issue #112)
-5. Run `node /server/db/create_test_data.js` to create test data (pass that script `--help` for options)
+5. Run `./bin/create_test_data.js` to create test data (pass that script `--help` for options)
+
 6. `npm start`
 7. You're up and running! surf to <http://localhost:8080> :surfer:
